V0 Membresia Funcionalidad BBDD (- modificar & actualizarlista)

+++ src/main/java/presentacion/MembresiaFrame.java
@@ -4,6 +4,7 @@ import dominio.ControlMembresia;
 import dominio.Membresia;
 import java.util.ArrayList;
 import javax.swing.DefaultListModel;
+import javax.swing.JOptionPane;
 
 /**
  * @author angel
@@ -108,16 +109,17 @@ public class MembresiaFrame extends javax.swing.JFrame {
     private void btnAnadirMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnAnadirMousePressed
         AnadirMembresia am = new AnadirMembresia();
         am.setVisible(true);
-        am.setLocationRelativeTo(null);
+        am.setLocationRelativeTo(null);        
     }//GEN-LAST:event_btnAnadirMousePressed
 
     private void btnEliminarMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnEliminarMousePressed
-        //poner codigo de confirmacion
-        if(cm.eliminarMembresia(listMembresias.getSelectedValue())){            
-            modeloLista.remove(listMembresias.getSelectedIndex());
-        }else{
-            //codigo no se ha podido elimnar
-        }               
+        int dialogButton = JOptionPane.YES_NO_OPTION;
+        int dialogResult = JOptionPane.showConfirmDialog (null, "Estas seguro que desea elimnar?","Warning",dialogButton);
+        if(dialogResult == JOptionPane.YES_OPTION){
+            if(cm.eliminarMembresia(listMembresias.getSelectedValue())){            
+                modeloLista.remove(listMembresias.getSelectedIndex());
+            }
+        }
     }//GEN-LAST:event_btnEliminarMousePressed
 
     private void listMembresiasMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_listMembresiasMouseClicked
@@ -143,43 +145,12 @@ public class MembresiaFrame extends javax.swing.JFrame {
         }
         return nombreMem;
     }      
-    
-    /**
-     * @param args the command line arguments
-     **/
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(MembresiaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(MembresiaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(MembresiaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(MembresiaFrame.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
 
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                new MembresiaFrame().setVisible(true);
-            }
-        });
+    public void actualizarLista(){
+        modeloLista.removeAllElements();
+        modeloLista.addAll(0, obtenerMembresias());
     }
-
+    
     private final ControlMembresia cm;    
     private DefaultListModel modeloLista;
